+++ experimental/torch_xla2/torch_xla2/ops/jaten.py
@@ -68,44 +68,6 @@ def _handle_int64_trig(self, func):
   return res
 
 
-@op(torch.ops._c10d_functional.all_gather_into_tensor)
-def _c10d_all_gather(input, group_size: int, group_name: str):
-  return jax.lax.all_gather(input, "torch_dist")
-
-
-# TODO(wcromar): move c10d ops since they're not actually aten
-@op(torch.ops._c10d_functional.all_reduce)
-def _c10d_all_reduce(self, reduceOp: str, group_name: str):
-  match reduceOp:
-    case 'sum':
-      res = jax.lax.psum(self, axis_name="torch_dist")
-    case 'avg':
-      res = jax.lax.pmean(self, axis_name="torch_dist")
-    case 'min':
-      res = jax.lax.pmin(self, axis_name="torch_dist")
-    case 'max':
-      res = jax.lax.pmax(self, axis_name="torch_dist")
-    case _:
-      raise RuntimeError(f"Reduce op {reduceOp} not implemented")
-  return res
-
-
-@op(torch.ops._c10d_functional.broadcast)
-def _c10d_broadcast(self, src: int, group_name: str):
-  masked = jnp.where(
-      jax.lax.axis_index("torch_dist") == src,
-      self,
-      jnp.zeros_like(self),
-  )
-  return jax.lax.psum(masked, "torch_dist")
-
-
-@op(torch.ops._c10d_functional.wait_tensor)
-def _c10d_wait_tensor(tensor):
-  # Async tensor is aleady `wait`ed by dispatcher
-  return tensor
-
-
 @op(
   torch.ops.aten.view_copy,
   torch.ops.aten.view,
